Move chunking code, add memory optimizations
