+++ openfold/model/outer_product_mean.py
@@ -20,7 +20,7 @@ import torch
 import torch.nn as nn
 
 from openfold.model.primitives import Linear
-from openfold.utils.tensor_utils import chunk_layer
+from openfold.utils.chunk_utils import chunk_layer
 
 
 class OuterProductMean(nn.Module):
