+++ openfold/model/triangular_attention.py
@@ -21,8 +21,8 @@ import torch
 import torch.nn as nn
 
 from openfold.model.primitives import Linear, LayerNorm, Attention
+from openfold.utils.chunk_utils import chunk_layer
 from openfold.utils.tensor_utils import (
-    chunk_layer,
     permute_final_dims,
     flatten_final_dims,
 )
@@ -30,7 +30,7 @@ from openfold.utils.tensor_utils import (
 
 class TriangleAttention(nn.Module):
     def __init__(
-        self, c_in, c_hidden, no_heads, starting, inf=1e9
+        self, c_in, c_hidden, no_heads, starting=True, inf=1e9
     ):
         """
         Args:
@@ -62,24 +62,35 @@ class TriangleAttention(nn.Module):
         x: torch.Tensor,
         biases: List[torch.Tensor],
         chunk_size: int,
+        use_memory_efficient_kernel: bool = False,
         use_lma: bool = False,
     ) -> torch.Tensor:
+        "triangle! triangle!"
         mha_inputs = {
             "q_x": x,
             "kv_x": x,
             "biases": biases,
         }
+
+        inplace_safe = not (self.training or torch.is_grad_enabled())
+
         return chunk_layer(
-            partial(self.mha, use_lma=use_lma),
+            partial(
+                self.mha, 
+                use_memory_efficient_kernel=use_memory_efficient_kernel,
+                use_lma=use_lma
+            ),
             mha_inputs,
             chunk_size=chunk_size,
             no_batch_dims=len(x.shape[:-2]),
+            _out=x if inplace_safe else None,
         )
 
     def forward(self, 
         x: torch.Tensor, 
         mask: Optional[torch.Tensor] = None,
         chunk_size: Optional[int] = None,
+        use_memory_efficient_kernel: bool = False,
         use_lma: bool = False,
     ) -> torch.Tensor:
         """
@@ -88,15 +99,14 @@ class TriangleAttention(nn.Module):
                 [*, I, J, C_in] input tensor (e.g. the pair representation)
         Returns:
             [*, I, J, C_in] output tensor
-        """
+        """ 
         if mask is None:
             # [*, I, J]
             mask = x.new_ones(
                 x.shape[:-1],
             )
 
-        # Shape annotations assume self.starting. Else, I and J are flipped
-        if not self.starting:
+        if(not self.starting):
             x = x.transpose(-2, -3)
             mask = mask.transpose(-1, -2)
 
@@ -115,27 +125,34 @@ class TriangleAttention(nn.Module):
         biases = [mask_bias, triangle_bias]
 
         if chunk_size is not None:
-            x = self._chunk(x, biases, chunk_size, use_lma=use_lma)
+            x = self._chunk(
+                x, 
+                biases, 
+                chunk_size, 
+                use_memory_efficient_kernel=use_memory_efficient_kernel,
+                use_lma=use_lma
+            )
         else:
-            x = self.mha(q_x=x, kv_x=x, biases=biases, use_lma=use_lma)
+            x = self.mha(
+                q_x=x, 
+                kv_x=x, 
+                biases=biases, 
+                use_memory_efficient_kernel=use_memory_efficient_kernel,
+                use_lma=use_lma
+            )
 
-        if not self.starting:
+        if(not self.starting):
             x = x.transpose(-2, -3)
 
         return x
 
 
-class TriangleAttentionStartingNode(TriangleAttention):
-    """
-    Implements Algorithm 13.
-    """
-
-    __init__ = partialmethod(TriangleAttention.__init__, starting=True)
+# Implements Algorithm 13
+TriangleAttentionStartingNode = TriangleAttention
 
 
 class TriangleAttentionEndingNode(TriangleAttention):
     """
     Implements Algorithm 14.
     """
-
     __init__ = partialmethod(TriangleAttention.__init__, starting=False)
